+++ packages/tx-utils/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- Drop support for CommonJS
+
 ## 0.2.5 - 2026-03-13
 
 ### Fixed